refactor: replace PageRenderManager with RenderManager and simplify rendering logic
